move helper functions to more specific files
